# zwf-app/src/_assets/api.yml
swagger: "2.0"
info:
  description: ZWF backend API
  version: "1.0.0"
  title: "ZWF API"
basePath: "/api/v1"
schemes:
  - http
  - https
...
paths:
...
  /task/deep/{deepLinkId}:
    get:
      operationId: getDeepLinkedTask
      # security:
      #   - authLoggedInUser: []
      parameters:
        - name: deepLinkId
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
...
  /task/{id}:
    get:
      operationId: getTask
      # security:
      #   - authLoggedInUser: []
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
  /task/{id}/status/{status}:
    post:
      operationId: changeTaskStatus
      # security:
      #   - authClient: []
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
        - name: status
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
  /task/{id}/tags:
    post:
      operationId: updateTaskTags
      # security:
      #   - authClient: []
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
  /task/{id}/rename:
    post:
      operationId: updateTaskName
      # security:
      #   - authClient: []
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
  /task/{id}/history:
    get:
      operationId: getTaskHistory
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
...
  /task/{id}/content/sse:
    get:
      operationId: subscribeTaskContent
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
...
  /task/{id}/fields:
    post:
      operationId: updateTaskFields
      # security:
      #   - authClient: []
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true        
      responses:
        200:
          description: OK      
  /task/{id}/field/value:
    post:
      operationId: saveTaskFieldValue
      # security:
      #   - authClient: []
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true        
      responses:
        200:
          description: OK      
  /task/field/{fieldId}/file:
    post:
      operationId: uploadTaskFieldFile
      parameters:
        - name: fieldId
          in: path
          type: string
          format: uuid
          required: true        
      responses:
        200:
          description: OK     
  /task/file/{fileId}:
    get:
      operationId: downloadTaskFile
      # security:
      #   - authLoggedInUser: []
      parameters:
        - name: fileId
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK           
  /task/file/{fileId}/sign:
    post:
      operationId: signTaskFile
      # security:
      #   - authLoggedInUser: []
      parameters:
        - name: fileId
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK              
...
  /task/{id}/assign:
    post:
      operationId: assignTask
      # security:
      #   - authClient: []
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
  /task/{id}/notify:
    post:
      operationId: notifyTask
      # security:
      #   - authClient: []
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK          
...